refactor(errors): separation of the errors inside an `error` folder + using Clone derive
