+++ src/controllers/remote_controller.rs
@@ -1,5 +1,6 @@
 mod model;
 
+use crate::error::gh_reqwestor_error::GhReqwestError;
 use crate::gh_reqwestor;
 use model::GithubOrganization;
 pub use model::RemotePortfolio;
@@ -19,7 +20,7 @@ const LIST_REMOTES_URL: &str = "https://api.github.com/user/orgs";
 pub async fn get_remotes(
     gh_auth_token: &str,
     _args: &Vec<String>,
-) -> Result<Vec<RemotePortfolio>, crate::error::GhReqwestError> {
+) -> Result<Vec<RemotePortfolio>, GhReqwestError> {
     let mut remotes: Vec<RemotePortfolio> = Vec::new();
     let user_orgs: Vec<GithubOrganization> =
         gh_reqwestor::get::<Vec<GithubOrganization>>(&LIST_REMOTES_URL, gh_auth_token).await?;
